feat: implement mutex concurrency handler

+++ src/infra/database/typeorm/typeorm-user-session.repository.ts
@@ -2,7 +2,9 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Session } from 'src/domain/entities/Session';
 import { Stream } from 'src/domain/entities/Stream';
+import { ResourceLocked } from 'src/domain/errors/resource-locked-error';
 import { UserSessionNotFoundError } from 'src/domain/errors/user-session-not-found-error';
+import { LockRepository } from 'src/domain/repository/lock-repository';
 import { UserSessionRepository } from 'src/domain/repository/user-session-repository';
 import { Repository } from 'typeorm';
 
@@ -13,13 +15,50 @@ export class TypeORMUserSessionRepository implements UserSessionRepository {
         private userSessionRepository: Repository<Session>,
         @InjectRepository(Stream)
         private streamRepository: Repository<Stream>,
+        private lockRepository: LockRepository
     ) { }
 
-    async save(userSession: Session): Promise<void> {
-        await this.userSessionRepository.save(userSession);
+    private async acquireLock(resourceId: string): Promise<boolean> {
+        const isLocked = await this.lockRepository.acquireLock(resourceId);
+
+        return isLocked;
     }
+
+    private async releaseLock(resourceId: string): Promise<void> {
+        await this.lockRepository.releaseLock(resourceId);
+    }
+
+    async save(userSession: Session): Promise<void> {
+        const isLocked = await this.acquireLock(userSession.getUserId());
+
+        if (isLocked) {
+            throw new ResourceLocked({
+                userSession: userSession.getUserId()
+            });
+        }
+
+        try {
+            await this.userSessionRepository.save(userSession);
+        }
+        finally {
+            await this.releaseLock(userSession.getUserId());
+        }
+    }
+
     async updateSession(session: Session): Promise<void> {
-        await this.userSessionRepository.save(session);
+        const isLocked = await this.acquireLock(session.getUserId());
+
+        if (isLocked) {
+            throw new ResourceLocked({
+                userSession: session.getUserId()
+            });
+        }
+
+        try {
+            await this.userSessionRepository.save(session);
+        } finally {
+            await this.releaseLock(session.getUserId());
+        }
     }
 
     async findByUserId(userId: string): Promise<Session | null> {
@@ -33,13 +72,26 @@ export class TypeORMUserSessionRepository implements UserSessionRepository {
 
         return sessionModel.toDomain();
     }
-    async deleteByUserId(session, streamId: string): Promise<void> {
+    async deleteByUserId(session: Session, streamId: string): Promise<void> {
+
         const streamSession = await this.userSessionRepository.findOne({ where: { userId: session.getUserId() } });
 
         if (!streamSession) {
             throw new UserSessionNotFoundError(session.getUserId());
         }
-        
-        await this.streamRepository.delete(streamId);
+
+        const isLocked = await this.acquireLock(session.getUserId());
+
+        if (isLocked) {
+            throw new ResourceLocked({
+                userSession: session.getUserId()
+            });
+        }
+
+        try {
+            await this.streamRepository.delete(streamId);
+        } finally {
+            await this.releaseLock(session.getUserId());
+        }
     }
 }
